more-work-on-structure
