+++ pages/_app.tsx
@@ -1,26 +1,25 @@
 import React from "react";
-import '../src/scss/globals.scss'
-import { NextPageWithLayout } from "../src/utils";
 import '../src/scss/globals.scss';
 import { SWRConfig } from "swr";
-import { $api } from "../src/services";
+import { AppProps } from "next/app";
+import { NextPageWithLayout } from "@utils/index";
+import { $api } from "@services/index";
 
-type Props = {
+interface Props extends AppProps {
   Component: NextPageWithLayout;
-  pageProps: any; // поки не знаю який тут тип прописувати
-}
+};
 
-const App = ({ Component, pageProps }: Props) => {
-  const getLayout = Component.getLayout ?? ((page) => page);
-  console.log(pageProps);
-  
+const App = (props: Props) => {
+  const { Component, pageProps } = props;
+
+  const getLayout = Component.getLayout ?? ((page) => page);  
 
   return (
     <>
       <SWRConfig
         value={{
           fetcher: (url) => $api.get(url).then((res) => res.data),
-          // revalidateIfStale: false,
+          revalidateIfStale: false,
           revalidateOnFocus: false,
           revalidateOnReconnect: false
         }}>
@@ -30,4 +29,4 @@ const App = ({ Component, pageProps }: Props) => {
   );
 };
 
-export default App
+export default App;
