+++ src/components/form/index.ts
@@ -1,0 +1,6 @@
+export * from './button';
+export * from './error';
+export * from './form';
+export * from './input';
+export * from './inputs-array';
+export * from './radio';
