+++ src/components/form/input/index.tsx
@@ -1,0 +1,1 @@
+export { };
