+++ src/components/form/inputs-array/index.tsx
@@ -1,0 +1,1 @@
+export { };
