+++ src/components/index.ts
@@ -2,3 +2,4 @@ export * from './cards';
 export * from './sections';
 export * from './common';
 export * from './ui';
+export * from './utils';
