+++ src/components/sections/index.ts
@@ -1,5 +1,2 @@
-export * from './contactForm';
-export * from './contactList';
 export * from './footer';
 export * from './header';
-export * from './toDoList';
