+++ src/components/ui/title/index.tsx
@@ -1,9 +1,11 @@
+import { FC } from "react";
+
 type Props = {
   title: string
 }
 
-const Title = ({ title }: Props) => (
+export const Title: FC<Props> = ({ title }) => {
+  return (
     <h1 className='page-title'>{title}</h1>
-)
-
-export default Title;
+  );
+};
