+++ src/components/utils/meta/index.tsx
@@ -1,16 +1,15 @@
+import { IMetadata } from "@utils/index";
 import Head from "next/head";
 import { FC } from "react";
-import { IMetadata } from "../../../utils";
 
 interface MetaProps {
   meta: IMetadata;
-}
+};
 
-export const Meta: FC<MetaProps> = ({meta}) => { 
+export const Meta: FC<MetaProps> = ({ meta }) => {
     return (
         <Head>
-            <title>{meta?.title || "Безсонне медіа"}</title>
-            <meta name="description" content={meta?.description} />
+            <title>{meta?.title || "Next.js"}</title>
         </Head>
     );
-}
+};
