+++ src/layouts/main/index.tsx
@@ -1,12 +1,12 @@
+import { Footer, Header } from '@components/index';
 import { FC, ReactNode } from 'react';
-import Footer from '../../components/sections/footer';
-import Header from '../../components/sections/header';
 import { SWRConfig } from "swr";
 
 type Props = {
   children?: ReactNode
   fallback?: any
-}
+};
+
 export const MainLayout: FC<Props> = (props) => {
   const { fallback = {}, children } = props;
