+++ src/pages/todos/index.tsx
@@ -1,11 +1,21 @@
-import Title from "../../components/ui/title";
-import ToDoList from "../../components/sections/toDoList";
+import { Title, ToDoItem } from "@components/index";
+import { useToDos } from "@hooks/index";
+import { ITodo } from "@utils/index";
+import { FC } from "react";
 
-const ToDos = () => (
-  <section>
-    <Title title='Todos List' />
-    <ToDoList/>
-  </section>
-);
+export const ToDos: FC = () => {
+  const todos: ITodo[] = useToDos();
 
-export default ToDos;
+  return (
+    <section>
+      <Title title='Todos List' />
+      <ul className='todo-list' >
+        {todos && todos.map((item) => (
+          <li key={item.id}>
+            <ToDoItem data={item} />
+          </li>
+        ))}
+      </ul>
+    </section>
+  );
+};
