--- src/scss/components/cards/toDoDetails/index.scss
@@ -1,11 +1,0 @@
-.todo-details {
-    margin-top: 2rem;
-
-    & > h2 {
-        font-size: 2.8rem;
-        
-        &:first-letter {
-            text-transform: uppercase;
-        }
-    }
-}
