+++ src/scss/components/pages/contacts/index.scss
@@ -4,4 +4,89 @@
         justify-content: space-around;
         padding-block: 3rem;
     }
+
+    .contact-form {
+        width: 45rem;
+        padding: 2rem;
+        background-color: var(--clr-default-200);
+        display: flex;
+        flex-direction: column;
+        gap: 1.5rem;
+
+        label {
+            margin-bottom: 1rem;
+        }
+    
+        .field {
+            position: relative;
+            display: flex;
+            flex-direction: column;
+
+            &-dob {
+                width: 100px;
+            }
+        }
+    
+        .input {
+            width: 100%;
+            height: 3rem;
+            margin-bottom: 0.3rem;
+        }
+    
+        .error {
+            position: absolute;
+            bottom: -1.4rem;
+            color: var(--clr-error);
+        }
+    
+        .button {
+            width: 20rem;
+            height: 3rem;
+            align-self: center;
+            cursor: pointer;
+        }
+    
+        .phones {
+            display: flex;
+            flex-direction: column;
+            gap: 1.7rem;
+    
+            &>div {
+                position: relative;
+                width: 100%;
+                display: flex;
+                justify-content: space-between;
+            }
+    
+            & input {
+                height: 3rem;
+                width: 25rem;
+            }
+    
+            & button {
+                width: 13rem;
+                height: 3rem;
+            }
+    
+            &>button {
+                width: 25rem;
+            }
+        }
+    }
+
+    .contact-list {
+        width: 40rem;
+        padding: 2rem;
+        background-color: var(--clr-default-200);
+
+        h2 {
+            margin-bottom: 1.5rem;
+        }
+    
+        ul {
+            display: flex;
+            flex-direction: column;
+            gap: 0.7rem;
+        }
+    }
 }
