+++ src/scss/components/pages/todo/index.scss
@@ -1,0 +1,11 @@
+.todo-details {
+    margin-top: 2rem;
+
+    &>h2 {
+        font-size: 2.8rem;
+
+        &:first-letter {
+            text-transform: uppercase;
+        }
+    }
+}
