+++ src/scss/components/pages/todos/index.scss
@@ -1,0 +1,7 @@
+.todo-list {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    width: 70rem;
+    margin: 0 auto;
+}
