--- src/scss/components/sections/contactForm/index.scss
@@ -1,64 +1,0 @@
-.contact-form {
-    width: 45rem;
-    padding: 2rem;
-    background-color: var(--clr-default-200);
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-
-    label {
-        margin-bottom: 1rem;
-    }
-
-    .field {
-        position: relative;
-        display: flex;
-        flex-direction: column;
-    }
-
-    .input {
-        width: 100%;
-        height: 3rem;
-        margin-bottom: 0.3rem;
-    }
-
-    .error {
-        position: absolute;
-        bottom: -1.4rem;
-        color: var(--clr-error);
-    }
-
-    .button {
-        width: 20rem;
-        height: 3rem;
-        align-self: center;
-        cursor: pointer;
-    }
-
-    .phones{
-        display: flex;
-        flex-direction: column;
-        gap: 1.7rem;
-        
-        &>div {
-            position: relative;
-            width: 100%;
-            display: flex;
-            justify-content: space-between;
-        }
-        
-        & input {
-            height: 3rem;
-            width: 25rem;
-        }
-        
-        & button {
-            width: 13rem;
-            height: 3rem;
-        }
-        
-        & > button {
-            width: 25rem;
-        }
-    }
-}
