--- src/scss/components/sections/contactList/index.scss
@@ -1,15 +1,0 @@
-.contact-list {
-    width: 40rem;
-    padding: 2rem;
-    background-color: var(--clr-default-200);
-
-    h2 {
-        margin-bottom: 1.5rem;
-    }
-    
-    ul {
-        display: flex;
-        flex-direction: column;
-        gap: 0.7rem;
-    }
-}
