--- src/scss/components/sections/toDoList/index.scss
@@ -1,7 +1,0 @@
-.todo-list {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    width: 70rem;
-    margin: 0 auto;
-}
