+++ src/utils/declarations/metadata/metadata.types.ts
@@ -1,4 +1,3 @@
 export interface IMetadata {
   title?: string;
-  description?: string;
 }
